Refatora estrutura do app dominio
