+++ dominio/login/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import login
+
+
+urlpatterns = [
+    path('login/', login, name="login"),
+]
