+++ dominio/tutela/serializers.py
@@ -71,17 +71,3 @@ class DetalheProcessosJuizoSerializer(serializers.Serializer):
     nr_acoes_propostas_60_dias = serializers.IntegerField(min_value=0)
     variacao_12_meses = serializers.FloatField()
     top_n = AcoesPromotoriaSerializer(many=True)
-
-
-class AlertasListaSerializer(serializers.Serializer):
-    sigla = serializers.CharField()
-    descricao = serializers.CharField()
-    doc_dk = serializers.IntegerField()
-    num_doc = serializers.CharField()
-    num_ext = serializers.CharField()
-    etiqueta = serializers.CharField()
-    classe_doc = serializers.CharField()
-    data_alerta = serializers.DateTimeField()
-    orgao = serializers.IntegerField()
-    classe_hier = serializers.CharField()
-    dias_passados = serializers.IntegerField()
