+++ dominio/tutela/tests/test_views.py
@@ -5,15 +5,15 @@ from django.conf import settings
 from django.test import TestCase
 from django.urls import reverse
 
-from dominio.views import DetalheProcessosJuizoView
+from dominio.tutela.views import DetalheProcessosJuizoView
 
-from .testconf import NoJWTTestCase, NoCacheTestCase
+from dominio.tests.testconf import NoJWTTestCase, NoCacheTestCase
 
 
 # Create your tests here.
 class DetalheAcervoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_acervo_variation_result(self, _run_query):
         _run_query.return_value = [
             (0, 'PROMO0', '30', '60', '-60.0'),
@@ -94,7 +94,7 @@ class DetalheAcervoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, expected_response)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_acervo_variation_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -108,8 +108,7 @@ class DetalheAcervoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
 
 class OutliersViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
-
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_outliers_result(self, _run_query):
         _run_query.return_value = [
             ('0', '10', '20', '100', '1000', '500', '300',
@@ -161,7 +160,7 @@ class OutliersViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, expected_response)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_outliers_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -175,8 +174,7 @@ class OutliersViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
 
 class SaidasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
-
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_saidas_result(self, _run_query):
         _run_query.return_value = \
             [
@@ -205,7 +203,7 @@ class SaidasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, expected_response)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_saidas_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -219,8 +217,7 @@ class SaidasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
 
 class EntradasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
-
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_entradas_result(self, _run_query):
         _run_query.return_value = \
             [
@@ -271,7 +268,7 @@ class EntradasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, expected_response)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_entradas_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -286,7 +283,7 @@ class EntradasViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
 class DetalheProcessosJuizoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_get_numero_acoes_propostas_pacote_atribuicao(self, _run_query):
         DetalheProcessosJuizoView\
             .get_numero_acoes_propostas_pacote_atribuicao(1)
@@ -311,7 +308,7 @@ class DetalheProcessosJuizoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
         _run_query.assert_called_once_with(expected_query, expected_parameters)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_detalhe_processos_result(self, _run_query):
         _run_query.side_effect = \
             [
@@ -337,7 +334,7 @@ class DetalheProcessosJuizoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, expected_response)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_detalhe_processos_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -351,8 +348,9 @@ class DetalheProcessosJuizoViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
 
 class ListaProcessosViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
-    @mock.patch('dominio.views.ListaProcessosView.PROCESSOS_SIZE')
-    @mock.patch('dominio.views.run_query')
+    # TODO: override settings
+    @mock.patch('dominio.tutela.views.ListaProcessosView.PROCESSOS_SIZE')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_lista_processos_result(self, _run_query, _PROCESSOS_SIZE):
         _PROCESSOS_SIZE.return_value = 1
 
@@ -419,7 +417,7 @@ class ListaProcessosViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
         self.assertEqual(response_1.data, expected_response_page_1)
         self.assertEqual(response_2.data, expected_response_page_2)
 
-    @mock.patch('dominio.views.run_query')
+    @mock.patch('dominio.tutela.views.run_query')
     def test_entradas_no_result(self, _run_query):
         _run_query.return_value = []
         response = self.client.get(reverse(
@@ -430,3 +428,63 @@ class ListaProcessosViewTest(NoJWTTestCase, NoCacheTestCase, TestCase):
 
         self.assertEqual(response.status_code, 404)
         self.assertEqual(response.data, expected_response)
+
+
+class TestTempoTramitacao(NoJWTTestCase, TestCase):
+    @mock.patch('dominio.tutela.views.run_query')
+    def test_correct_response(self, _run_query):
+        expected = {
+            "id_orgao": 12345,
+            "media_orgao": 10.1243,
+            "minimo_orgao": 0,
+            "maximo_orgao": 100,
+            "mediana_orgao": 10.2312,
+            "media_pacote": 11.4352,
+            "minimo_pacote": 0,
+            "maximo_pacote": 200,
+            "mediana_pacote": 56.3124,
+            "media_pacote_t1": 45.343,
+            "minimo_pacote_t1": 12,
+            "maximo_pacote_t1": 533,
+            "mediana_pacote_t1": 343.324,
+            "media_orgao_t1": 344.12,
+            "minimo_orgao_t1": 12,
+            "maximo_orgao_t1": 5023,
+            "mediana_orgao_t1": 2421.1223,
+            "media_pacote_t2": 343.1254,
+            "minimo_pacote_t2": 48,
+            "maximo_pacote_t2": 2335,
+            "mediana_pacote_t2": 7623.1224,
+            "media_orgao_t2": 43224.1132,
+            "minimo_orgao_t2": 432,
+            "maximo_orgao_t2": 1324,
+            "mediana_orgao_t2": 2242.3232
+        }
+        _run_query.return_value = [expected.values()]
+        url = reverse("dominio:tempo-tramitacao", args=("1234", ))
+        resp = self.client.get(url)
+
+        self.assertEqual(resp.status_code, 200)
+        self.assertEqual(resp.data, expected)
+
+
+class TestNumeroDesarquivamentos(NoJWTTestCase, TestCase):
+    @mock.patch("dominio.tutela.views.connections")
+    def test_correct_response(self, _connections):
+        cursor_mock = mock.MagicMock()
+        cursor_mock.execute.return_value.fetchall.return_value\
+            = [("nr_mp_1", 1), ("nr_mp_2", 2)]
+        conn_mock = mock.MagicMock()
+        conn_mock.cursor.return_value.__enter__.return_value = cursor_mock
+
+        _connections.__getitem__.return_value = conn_mock
+        url = reverse("dominio:desarquivamentos", args=("12345",))
+
+        resp = self.client.get(url)
+        expected = [
+            {"numero_mprj": "nr_mp_1", "qtd_desarq": 1},
+            {"numero_mprj": "nr_mp_2", "qtd_desarq": 2},
+        ]
+
+        self.assertEqual(resp.status_code, 200)
+        self.assertEqual(resp.data, expected)
